Use ruby-lsp allowing it to use rubocop for formatting
Configuring rubocop in conform.nvim woudl not work - just gave timeouts
Configuring ruby-fmt in conform worked, but used different rules to the ruby-lsp/rubocop diagnostics - not fun!
Instead, ignore conform for formatting and just allow it to fall back to ruby-lsp, which is configured to  use rubocop for formatting.
Result: Nice inline diagnostics from rub-lsp and format on save to appease rubocop. Hurrah!